moves GameLevelNavigationLink to its own file
