+++ Sudoku/ContentView.swift
@@ -8,22 +8,6 @@
 
 import SwiftUI
 
-struct GameLevelNavigationLink: View {
-    
-    let level: Difficulty.Level
-    
-    var body: some View {
-        NavigationLink {
-            GameView(workingGrid: GridValues(startingGrid: GridFactory.randomGridForDifficulty(level: level)), 
-                     viewModel: GameViewModel(difficulty: level))
-        } label: {
-            Text(level.rawValue)
-                .font(.system(.headline, design: .rounded))
-        }
-        .dynamicButtonStyle(backgroundColor: Color.blue.opacity(0.2))
-    }
-}
-
 struct ContentView: View {
     
     private let viewModel = ContentViewModel()
